Update rails_helper to require files automatically
You'll normally find this in the spec_helper.rb, but when using rails and a rails_helper.rb Rspec puts its default require all support files logic here.

We tweak it slightly to avoid pulling in simplecov twice (its required at the top of spec_helper, which in turn is required at the top of rails_helper), and this blob of code is pretty standard across all our projects.