Alter namespace for project
This change updates the namespace throughout the project to match the convention we have used in our `DefraRuby` gems.

In those we have the main namespace of `DefraRuby` and then the features specific namespace, for example `Address`, `Aws`, `Validations`, and `Alert`.

We want this engine to operate in the same way, so this change goes through the project and updates the namespace from `DefraRubyMocks` to `DefraRuby::Mocks`.

It also adds a temporary route just so we could confirm that this was working.