Add companies house mock
All our current services (see https://github.com/DEFRA/ruby-services-team/tree/master/services) feature a basic form of Companies house integration.

When performance testing our services we don't want to be responsible for subjecting the actual Companies House API to unnecessary load, hence we need some way to mock it.

This change adds our first mock, which is for Companies house. Essentially we will store a list of recorded responses which we will return if a matching company number is provided. Else we'll return no result found if a company number is used which does not match one of our recordings.