Respond with actual JSON for not found
When the real Companies House API responds with a 404, it returns a static piece of JSON. It's only a small piece of JSON, and it does not change based on the request. So this change adds this to the project so we can respond in the way way the real API does.