Merge namespaces back together
In a previous commit we split the namespace up from `DefraRubyMocks` to `DefraRuby::Mocks`.

Our intention was to follow the convention of our gems, where the namespace will be `DefraRuby::Address` for example, but the gem name is `defra_ruby_address`.

However there did not seem to be a way to do this when working with an engine, that would mean it would load automatically when mounted into an app.

We could only make it work using an additional require statement in the host apps `application.rb`.

So seeing as the naming was just a nicety to make it match with our other projects, we've reverted back the change to ensure the engine mounts and behaves in the same way as our other engines.