Add companies house mock (#4)
All our current services (see https://github.com/DEFRA/ruby-services-team/tree/master/services) feature a basic form of Companies house integration.

When performance testing our services we don't want to be responsible for subjecting the actual Companies House API to unnecessary load, hence we need some way to mock it.

This change adds our first mock, which is for Companies house. When you ping the Companies House API they will return you a whole heap of data. In our services, though we are currently only interested in the `company_status`.

The old Java services app maintained a store or previously recorded responses, however, this meant you needed to know which company numbers we had stored, but that was a lot of code.

The solution we have gone within this project is to just say, as long as the company number is in a valid format, we'll just return the bit of JSON the services need and say the status is `active`.

To allow for testing, we do still mimic what happens if you query with an invalid number. We'll also provide a list of 'special' numbers that if used will provide one of the other statuses Companies House may return.