Add configuration to the project
When we implemented mocks before, we have also provided a way to both enable the mocks (or disable them) but also implement a delay in the response.

By default we don't want the mocks enabled to protect against them accidently being used in production. And because we are not crunching through some massive database or carrying out any complex logic, our mock responses are likely to be far quicker than the real ones. So we also provide a means of introducing a configurable delay in the mock response.