Update to handle strings and booleans in config
Spotted that we may have an issue with our config. The use case we expect will be something like

```ruby
DefraRubyMocks.configure do |config|
  config.enabled = ENV["USE_MOCKS"] || false
end
```

`ENV["USE_MOCKS"]` will return a string, even if it is set to `"true"`. Our problem would be if it was set to `"false"`. That would be interpreted as 'truthy' and could cause the mocks to be enabled when the user does not want them to be.

So we have implemented our own setter to handle the value being passed in as either a string or a boolean. And because we have extended the logic, we felt it cleaner and clearer if we moved the class to its own file with a spec to cover this usage.