Refactor specs and handle strings and booleans (#7)
This change covers some refactoring of the specs, but also updates the configuration to handle the host app passing the value to `configuration.enable` as both a boolean and a string.

It also adds a test to check that disabling the mocks does work.

Further notes

Updating the configuration was because of something we spotted. The typical use case we expect will be something like

```ruby
DefraRubyMocks.configure do |config|
  config.enabled = ENV["USE_MOCKS"] || false
end
```

`ENV["USE_MOCKS"]` will return a string, even if it is set to `"true"`. Our problem would be if it was set to `"false"`. That would be interpreted as 'truthy' and could cause the mocks to be enabled when the user does not want them to be.

We could leave the host app to handle this, and just be explicit in our documentation that the value must be a boolean and not a string. But we felt it was something better handled here in one place.

So we have implemented our own setter to handle the value being passed in as either a string or a boolean. And because we have extended the logic, we felt it cleaner and clearer if we moved the class to its own file with a spec to cover this usage.