Fix 'config.enabled'. Should be 'enable'
When double checking the documentation for where we are at, we spotted that it referred to the field as `enable` whereas the code had it as `enabled`.

On review `enabled` would be better suited to a method called `enabled?` so we opted for the name used in the documentation and updated the code accordingly.