Fix Configuration.delay to handle strings
A bit like changes we made to `enable` to handle values being passed in as strings rather than integers, we should have spotted that the same will apply to `delay`.

Again we can expect the same usage patterns for setting the config

```ruby
DefraRubyMocks::configure do |config|
  config.enable = ENV["ENABLE_MOCKS"] || false
  config.enable = ENV["MOCK_DELAY"] || 1000
end
```

So there is a good chance the value of the env var will just be chucked straight across rather than the host app converting it first to an integer.