Update initial route
Using the implementation in waste-carriers-service as a guide, we need to 2 route paths. One to send the initial worldpay request to, and one to redirect the user to.

In Worldpay (using its current paths) these are named `payment-service` and `dispatcher` and so the WCR service has matching named endpoints.

This change just updates our basic hand cranked initial implementation for the payment-service to be named as such.