Add config and base logic for initial response
Starting to build up the functionality. We need the host app to tell us some info in order to generate valid responses, for example where exactly to redirect the 'user' to (the mock engine does not know where or how its been mounted, so needs to be told!)

Plus we add an initial function to start generating a 10 digit ID that worldpay seems to use.