Read response attributes from request
Some of the values we send back in the initial response need to come from the request sent through by the calling app.

Again this is an initial implementation of the controller parsing the request body to get those attributes, and ensuring that are in the response we send back.