Refactor tests in worldpay response spec
When about to start adding more tests to this spec we realised we'd need a bunch of mocks setup in order for things to work.

With them in place, we might as well use them for the existing tests, hence this first step is to refactor what we already have before adding the new unit tests.