Add worlpday mocking functionality (#11)
The [Waste carriers registration service](https://github.com/DEFRA/ruby-services-team/tree/master/services/wcr) currently integrates with Worldpay to take payments.

To support performance testing we created a 'mock' version of this in the [waste-carriers-service](https://github.com/DEFRA/waste-carriers-service). However that Java service is soon to be retired hence the incentive to create this engine.

This change re-implements that Worldpay mock.