Add mock Worldpay refund functionality
https://eaflood.atlassian.net/browse/RUBY-818

Though the mocking we are replacing in [waste-carriers-service](https://github.com/DEFRA/waste-carriers-service) never handled refund requests to Worldpay, we have chosen to add it here.

The primary need for mocking the external services is for performance testing, and when we do that we tend to focus on the new and renew registration journeys in the Waste Carriers Service.

However it's easy to forget mocking is enabled in an environment, and so it can cause confusion when something suddenly stops working leading to wasted time and effort.

So as it's only a minor enhancement, this change adds support for mocking refund requests to Worldpay. This will also mean we 'mock' all the interactions Waste Carriers has with Worldpay.