Add ability for Worldpay to reject payment
It was highlighted by the team's QA that some of our automated acceptance tests replicate Worldpay rejecting a payment in order to see how the system handles it.

Currently the mocks can't do this, so we were asked what could be done instead.

Our response was to say they can (will!) do this.