Fix generated MAC for refused payments
Spotted whilst working on a new enhancement (PR #32) that if are asked to return a `REFUSED` payment response that we are not actually generating the MAC correctly.

The MAC needs to include the status when generated but we had it hard coded to always be `AUTHORISED`. This means if the registration in question has the term 'reject' in its company name, which means we are being asked to mock a `REFUSED` response, the MAC we generate won't match what the app expects because it will be expecting one where the status is `REFUSED`.

We do already work out the status correctly. We're just not getting it to the method where we generate the MAC.

This change fixes that.