Add support for copy card orders in WCR (#32)
https://eaflood.atlassian.net/browse/RUBY-1044

The old [waste-carriers-frontend](https://github.com/DEFRA/waste-carriers-frontend) had a copy cards feature that simply piggybacked on the registration and used the same process as used when paying for a new registration.

Our Worldpay mock handles this, plus dealing with renewals via transient registrations collection in the database. That version of ordering copy cards is now defunct and has been replaced by a new version in the [https://github.com/DEFRA/waste-carriers-back-office).

It creates a separate transient registration of type `WasteCarriersEngine::OrderCopyCardsRegistration` which only includes a few details from the original registration. Unfortunately `company_name` is not one of them.

Our `WorldpayResponseService` uses the company name to determine if we should mock Worldpay `REFUSE`ing the payment. We look in the company name for the term `reject` and if found `REFUSE` the payment.

We could have amended the code so that if what we are dealing with is a `OrderCopyCardsRegistration` we always return a successful response. But instead, we decided to keep the option to simulate Worldpay refusing a copy card order payment. In this way, our QA has the scope and flexibility to test what happens in these scenarios.

So this change updates the `WorldpayResponseService` to handle tracking down the company name for copy card orders so they can still be refused and possibly other result types in the future.

* Other notes

Applied a pattern used in our apps of referring to the instance as a 'resource' rather than a 'registration' to make it clearer we may not actually be referring to a `Registration` instance.

Using @resource.is_a?` works fine when loaded in a waste carriers app. But we have written the engine to not depend on that being the case. This is why any tests reliant on the `WorldpayResponseService` were failing. Switching to simply checking the class name got around this problem.