Update WP response service to use resource service
This update our `WorldpayResponseService` having added the `WorldpayResourceService`. We're able to remove a bunch of code that's now covered by stuff in the new class.

We're also changing how we return a result. Going forward we'll use a struct that has a number of properties (and is easier to work with than a raw hash). There is some forward thinking here to make debugging and logging contain more info which should help track down issues in the future.