Update worldpay controller to handle new stuck page
In order to mimic a user getting stuck at worldpay and not returning to our service, we need a means to keep them on 'Worldpay'.

So now if the request is for a company with the term 'stuck' in their company name we'll simply render a new 'Stuck!' page when the request comes in. This means anyone testing has some idea of what is going on. But also that no response gets back to the requesting app, mocking what we believe happens in the real world.