Add support for 'stuck' Worldpay payments (#34)
https://eaflood.atlassian.net/browse/RUBY-1035

This isn't an official Worldpay response. This is in order to support recreating the scenario where a user is redirected to Worldpay but is then not redirected back.

We refer to the registration as being 'stuck at payment' and there is a feature built into the back-office that allows NCCC to move the registration back to the payment summary screen.

Our QA would like to cover this functionality with an automated acceptance test. So to support this we're adding the ability to tell the mock not to respond when the company name contains the term 'stuck'. This is similar to our existing [refuse payment](https://github.com/DEFRA/defra-ruby-mocks#refused-payments) functionality.

Included in this change is some refactoring of the `WorldpayResponseService`, breaking it up into 2 classes and providing more info in the result which we can then use in the controller.