Implement mock payment screens
We are being called on more and more to mock specific responses from Worldpay. Currently our process for handling this is to rely on checking the company number of the registration and responding according to that.

That's fine, but it does require extra complexity in trying to locate the registration in the database. The lack of any 'mock' pages also means our acceptance tests have to include special logic to handle the fact they may or may not see pages from Worldpay.

So this change is all about first adding pages that can actually be interacted with to mock Worldpay's own (including matching ID's for the html elements).

We can then build on this and like Worldpay's own test environment, use the values entered into our mock page to control how we respond, rather than having to set company names to include specific values.