Complete implementation of new stuck page
This completes moving the existing stuck page into its own controller. We've had to do a few things to conserve the current functionality.

The main thing is getting our params across to it for debug purposes. We've chosen to simply load stuff into a cookie for now (some of our values are pretty long!)

This means taking the struct returned by the response, converting it to a hash, and then converting that to JSON. We then drop that in the cookie and read it in the stuck action.

We've also tweaked the stuck page itself. Because you can now request it directly, we've added some logic so it can handle this and when rendered be distinctive that the request was direct.