Refactor worldpay controller (#36)
This is the first stage in refactoring the Worldpay mocking to actually present a page to users rather than automatically return.

We are being called on more and more to mock specific responses from Worldpay. Currently, our process for handling this is to rely on checking the company number of the registration and responding according to that.

That's fine, but it does require extra complexity in trying to locate the registration in the database. The lack of any 'mock' pages also means our acceptance tests have to include special logic to handle the fact they may or may not see pages from Worldpay.

To get ready for actually rendering a mock UI this change covers splitting the existing Worldpay controller into those actions which are just expected to provide a response with no UI, and those that are expected to render a page.

At this time that's just the stuck page, but it will give us the basis to add further pages in later changes.

> Essentially this is quite a big shift in functionality and we don't want to do it in one massive commit!