Add mocking of pending Worldpay payments
https://eaflood.atlassian.net/browse/RUBY-1045

This change adds the ability to 'mock' a pening Worldpay payment to the engine.

There are times when the Worldpay payment completes and it returns a `PENDING` response but this does not actually mean payment has been taken. We believe this happens when further checks need to be made before transfer of funds can be authorised by the bank.

We cater for this in both the renewal and new egistration journeys so our QA @andrewhick would also like to provide coverage via our acceptance tests.