Add mocking of pending Worldpay payments (#38)
https://eaflood.atlassian.net/browse/RUBY-1045

This change adds the ability to 'mock' a pending Worldpay payment to the engine.

There are times when the Worldpay payment completes and it returns a `PENDING` response but this does not actually mean payment has been taken. We believe this happens when further checks need to be made before the transfer of funds can be authorised by the bank.

We cater for this in both the renewal and new registration journeys so our QA @andrewhick would also like to provide coverage via our acceptance tests.

Note. In making this change we amended the signature of the WorldpayResponseService. But our Worldpay request tests still all passed. So this also includes some refactoring of the tests to ensure the double more closely matches the service it's mocking.