Add mocking of cancelled Worldpay payments
This change adds the ability to 'mock' a cancelled Worldpay payment to the engine.

We generally see this when a user chooses to cancel the payment process whilst on the Worldpay site. Worldpay will redirect them back to us with a response that includes this payment status.

We cater for this in both the renewal and new registration journeys so our QA @andrewhick would also like to provide coverage via our acceptance tests.