Add mocking of cancelled Worldpay payments (#39)
This change adds the ability to 'mock' a cancelled Worldpay payment to the engine.

We generally see this when a user chooses to cancel the payment process whilst on the Worldpay site. Worldpay will redirect them back to us with a response that includes this payment status.

We cater for this in both the renewal and new registration journeys so our QA @andrewhick would also like to provide coverage via our acceptance tests.

Note. The MAC for a cancelled payment does not include the payment status (appears to be the only case where it doesn't). Hence some necessary changes to handle that.